# admin-ui/api/generated/models/AdminCreateUserRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AdminCreateUserRequest = {
    username: string;
    password: string;
    full_name?: (string | null);
    is_admin?: boolean;
    is_active?: boolean;
};