# admin-ui/api/generated/models/SaleItemCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SaleItemCreate = {
    master_card_id: string;
    quantity: (number | string);
    unit_sale_price_rub: (number | string);
    fee_rub?: (number | string);
    extra_cost_rub?: (number | string);
    source_offer_id?: (string | null);
};